+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 0.2.0 (2026-05-06)
+
+- feat: support progress info panel style（theme or custom inline-style）
+- optimize: close panel event change to DOMContentLoaded（old is load）
+
 ## 0.1.0 (2026-05-06)
 
 - feat: Init base progress func